+++ pandas_gbq/gbq.py
@@ -19,6 +19,8 @@ import numpy as np
 if typing.TYPE_CHECKING:  # pragma: NO COVER
     import pandas
 
+import pandas_gbq.constants
+import pandas_gbq.exceptions
 from pandas_gbq.exceptions import GenericGBQException, QueryTimeout
 from pandas_gbq.features import FEATURES
 import pandas_gbq.query
@@ -266,8 +268,8 @@ class GbqConnector(object):
         client_secret=None,
     ):
         global context
-        from google.api_core.exceptions import GoogleAPIError
-        from google.api_core.exceptions import ClientError
+        from google.api_core.exceptions import ClientError, GoogleAPIError
+
         from pandas_gbq import auth
 
         self.http_error = (ClientError, GoogleAPIError)
@@ -411,7 +413,20 @@ class GbqConnector(object):
         timeout_ms = job_config_dict.get("jobTimeoutMs") or job_config_dict[
             "query"
         ].get("timeoutMs")
-        timeout_ms = int(timeout_ms) if timeout_ms else None
+
+        if timeout_ms:
+            timeout_ms = int(timeout_ms)
+            # Having too small a timeout_ms results in individual
+            # API calls timing out before they can finish.
+            # ~300 milliseconds is rule of thumb for bare minimum
+            # latency from the BigQuery API.
+            minimum_latency = 400
+            if timeout_ms < minimum_latency:
+                raise QueryTimeout(
+                    f"Query timeout must be at least 400 milliseconds: timeout_ms equals {timeout_ms}."
+                )
+        else:
+            timeout_ms = None
 
         self._start_timer()
         job_config = bigquery.QueryJobConfig.from_api_repr(job_config_dict)
@@ -464,6 +479,34 @@ class GbqConnector(object):
         create_bqstorage_client = self.use_bqstorage_api
         if max_results is not None:
             create_bqstorage_client = False
+
+        # If we're downloading a large table, BigQuery DataFrames might be a
+        # better fit. Not all code paths will populate rows_iter._table, but
+        # if it's not populated that means we are working with a small result
+        # set.
+        if (table_ref := getattr(rows_iter, "_table", None)) is not None:
+            table = self.client.get_table(table_ref)
+            if (
+                isinstance((num_bytes := table.num_bytes), int)
+                and num_bytes > pandas_gbq.constants.BYTES_TO_RECOMMEND_BIGFRAMES
+            ):
+                num_gib = num_bytes / pandas_gbq.constants.BYTES_IN_GIB
+                warnings.warn(
+                    f"Recommendation: Your results are {num_gib:.1f} GiB. "
+                    "Consider using BigQuery DataFrames (https://bit.ly/bigframes-intro)"
+                    "to process large results with pandas compatible APIs with transparent SQL "
+                    "pushdown to BigQuery engine. This provides an opportunity to save on costs "
+                    "and improve performance. "
+                    "Please reach out to bigframes-feedback@google.com with any "
+                    "questions or concerns. To disable this message, run "
+                    "warnings.simplefilter('ignore', category=pandas_gbq.exceptions.LargeResultsWarning)",
+                    category=pandas_gbq.exceptions.LargeResultsWarning,
+                    # user's code
+                    # -> read_gbq
+                    # -> run_query
+                    # -> download_results
+                    stacklevel=4,
+                )
 
         try:
             schema_fields = [field.to_api_repr() for field in rows_iter.schema]
@@ -650,17 +693,24 @@ def read_gbq(
     *,
     col_order=None,
 ):
-    r"""Load data from Google BigQuery using google-cloud-python
+    r"""Read data from Google BigQuery to a pandas DataFrame.
 
-    The main method a user calls to execute a Query in Google BigQuery
-    and read results into a pandas DataFrame.
-
-    This method uses the Google Cloud client library to make requests to
-    Google BigQuery, documented `here
-    <https://googleapis.dev/python/bigquery/latest/index.html>`__.
+    Run a SQL query in BigQuery or read directly from a table
+    the `Python client library for BigQuery
+    <https://cloud.google.com/python/docs/reference/bigquery/latest/index.html>`__
+    and for `BigQuery Storage
+    <https://cloud.google.com/python/docs/reference/bigquerystorage/latest>`__
+    to make API requests.
 
     See the :ref:`How to authenticate with Google BigQuery <authentication>`
     guide for authentication instructions.
+
+    .. note::
+        Consider using `BigQuery DataFrames
+        <https://cloud.google.com/bigquery/docs/dataframes-quickstart>`__ to
+        process large results with pandas compatible APIs that run in the
+        BigQuery SQL query engine. This provides an opportunity to save on
+        costs and improve performance.
 
     Parameters
     ----------
@@ -1037,12 +1087,7 @@ def to_gbq(
         )
 
     if api_method == "default":
-        # Avoid using parquet if pandas doesn't support lossless conversions to
-        # parquet timestamp. See: https://stackoverflow.com/a/69758676/101923
-        if FEATURES.pandas_has_parquet_with_lossless_timestamp:
-            api_method = "load_parquet"
-        else:
-            api_method = "load_csv"
+        api_method = "load_parquet"
 
     if chunksize is not None:
         if api_method == "load_parquet":
@@ -1199,8 +1244,7 @@ class _Table(GbqConnector):
 
     def _table_ref(self, table_id):
         """Return a BigQuery client library table reference"""
-        from google.cloud.bigquery import DatasetReference
-        from google.cloud.bigquery import TableReference
+        from google.cloud.bigquery import DatasetReference, TableReference
 
         return TableReference(
             DatasetReference(self.project_id, self.dataset_id), table_id
@@ -1241,9 +1285,7 @@ class _Table(GbqConnector):
             Use the generate_bq_schema to generate your table schema from a
             dataframe.
         """
-        from google.cloud.bigquery import DatasetReference
-        from google.cloud.bigquery import Table
-        from google.cloud.bigquery import TableReference
+        from google.cloud.bigquery import DatasetReference, Table, TableReference
 
         if self.exists(table_id):
             raise TableCreationError("Table {0} already exists".format(table_id))